+++ python/semantic_kernel/template_engine/blocks/block.py
@@ -1,34 +1,19 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-from abc import ABC, abstractmethod
 from logging import Logger
 from typing import Optional, Tuple
 
-from semantic_kernel.orchestration.context_variables import ContextVariables
-from semantic_kernel.orchestration.sk_context import SKContext
-from semantic_kernel.template_engine.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
+from semantic_kernel.utils.null_logger import NullLogger
 
 
-class Block(ABC):
-    _type: BlockTypes
-    _content: str
-    _log: Logger
-
-    def __init__(self, block_type: BlockTypes, content: str, log: Logger) -> None:
-        self._type = block_type
-        self._content = content
-        self._log = log
-
-    async def render_code_async(self, context: SKContext) -> str:
-        raise NotImplementedError("This block does not support code execution")
-
-    @abstractmethod
-    def is_valid(self) -> Tuple[bool, str]:
-        pass
-
-    @abstractmethod
-    def render(self, variables: Optional[ContextVariables]) -> str:
-        pass
+class Block:
+    def __init__(
+        self, content: Optional[str] = None, log: Optional[Logger] = NullLogger
+    ) -> None:
+        self._content = content or ""
+        self._log = log or NullLogger()
+        self._type = BlockTypes.UNDEFINED
 
     @property
     def type(self) -> BlockTypes:
@@ -37,3 +22,10 @@ class Block(ABC):
     @property
     def content(self) -> str:
         return self._content
+
+    @property
+    def log(self) -> Logger:
+        return self._log
+
+    def is_valid(self) -> Tuple[bool, str]:
+        raise NotImplementedError("Subclasses must implement this method.")
