+++ python/semantic_kernel/template_engine/blocks/block_types.py
@@ -1,10 +1,12 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-from enum import Enum
+from enum import Enum, auto
 
 
 class BlockTypes(Enum):
-    Undefined = 0
-    Text = 1
-    Code = 2
-    Variable = 3
+    UNDEFINED = auto()
+    TEXT = auto()
+    CODE = auto()
+    VARIABLE = auto()
+    VALUE = auto()
+    FUNCTION_ID = auto()
