+++ python/semantic_kernel/template_engine/blocks/function_id_block.py
