+++ python/semantic_kernel/template_engine/blocks/symbols.py
