+++ python/semantic_kernel/template_engine/protocols/prompt_templating_engine.py
