+++ python/semantic_kernel/template_engine/protocols/text_renderer.py
