--- python/semantic_kernel/template_engine_v2/blocks/block_types.py
@@ -1,12 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-from enum import Enum, auto
-
-
-class BlockTypes(Enum):
-    UNDEFINED = auto()
-    TEXT = auto()
-    CODE = auto()
-    VARIABLE = auto()
-    VALUE = auto()
-    FUNCTION_ID = auto()
